backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
bcrypt
pyjwt
pytest
pytest-asyncio
httpx
email-validator
psycopg2-binary
